// src/services/rental-api/generated/model/ctl-setting.ts
/* tslint:disable */
/* eslint-disable */
/**
 * pl-car-rental-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CtlSetting
 */
export interface CtlSetting {
    /**
     * 
     * @type {string}
     * @memberof CtlSetting
     */
    'key'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlSetting
     */
    'value'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlSetting
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlSetting
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlSetting
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlSetting
     */
    'updateTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlSetting
     */
    'description'?: string | null;
}